Changes
So, the original firebase.js was causing merging problems. So I added the key for our Firebase into index.js. I added the login.jsx and the signup.jsx. In the login.jsx it should handle logins with email and password only and so far restricted to ucla emails and it has both @g.ucla.edu and @ucla.edu to account for new students.